fix for lab

+++ week-7-networking/31-monday-JSON/labs/gson-lab/starter-code/GSON-lab/app/src/main/java/com/example/hollis/gson_example/Item.java
@@ -11,9 +11,9 @@ import com.google.gson.annotations.SerializedName;
  */
 public class Item {
     private String title;
-    private long lat;
+    private double lat;
     @SerializedName("long")
-    private long longitude;
+    private double longitude;
     private String link;
     private String pubDate;
     private Condition condition;
